+++ sdk/network/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/implementation/SecurityUserRuleCollectionsClientImpl.java
@@ -167,7 +167,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
             return Mono
                 .error(new IllegalArgumentException("Parameter configurationName is required and cannot be null."));
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.list(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(),
@@ -218,7 +218,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
             return Mono
                 .error(new IllegalArgumentException("Parameter configurationName is required and cannot be null."));
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service
@@ -380,7 +380,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
             return Mono
                 .error(new IllegalArgumentException("Parameter ruleCollectionName is required and cannot be null."));
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.get(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(),
@@ -429,7 +429,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
             return Mono
                 .error(new IllegalArgumentException("Parameter ruleCollectionName is required and cannot be null."));
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service.get(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(), resourceGroupName,
@@ -542,7 +542,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
         } else {
             securityUserRuleCollection.validate();
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.createOrUpdate(this.client.getEndpoint(), apiVersion,
@@ -600,7 +600,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
         } else {
             securityUserRuleCollection.validate();
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service.createOrUpdate(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(),
@@ -713,7 +713,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
             return Mono
                 .error(new IllegalArgumentException("Parameter ruleCollectionName is required and cannot be null."));
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.delete(this.client.getEndpoint(), apiVersion,
@@ -765,7 +765,7 @@ public final class SecurityUserRuleCollectionsClientImpl implements SecurityUser
             return Mono
                 .error(new IllegalArgumentException("Parameter ruleCollectionName is required and cannot be null."));
         }
-        final String apiVersion = "2024-10-01";
+        final String apiVersion = "2025-01-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service.delete(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(), resourceGroupName,