+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/SearchAsyncClient.java
@@ -1321,13 +1321,7 @@ public final class SearchAsyncClient {
             .searchPostWithResponseAsync(requestToUse, querySourceAuthorization, enableElevatedRead, null, context)
             .onErrorMap(MappingUtils::exceptionMapper)
             .map(response -> {
-                SearchDocumentsResult result = response.getValue();
-
-                SearchPagedResponse page
-                    = new SearchPagedResponse(new SimpleResponse<>(response, getSearchResults(result, serializer)),
-                        createContinuationToken(result, serviceVersion), result.getFacets(), result.getCount(),
-                        result.getCoverage(), result.getAnswers(), result.getSemanticPartialResponseReason(),
-                        result.getSemanticPartialResponseType());
+                SearchPagedResponse page = mapToSearchPagedResponse(response, serializer, serviceVersion);
                 if (continuationToken == null) {
                     firstPageResponseWrapper.setFirstPageResponse(page);
                 }
@@ -1345,6 +1339,16 @@ public final class SearchAsyncClient {
     static String createContinuationToken(SearchDocumentsResult result, ServiceVersion serviceVersion) {
         return SearchContinuationToken.serializeToken(serviceVersion.getVersion(), result.getNextLink(),
             result.getNextPageParameters());
+    }
+
+    static SearchPagedResponse mapToSearchPagedResponse(Response<SearchDocumentsResult> response,
+        JsonSerializer serializer, SearchServiceVersion serviceVersion) {
+        SearchDocumentsResult result = response.getValue();
+        return new SearchPagedResponse(new SimpleResponse<>(response, getSearchResults(result, serializer)),
+            createContinuationToken(result, serviceVersion), result.getFacets(), result.getCount(),
+            result.getCoverage(), result.getAnswers(), result.getSemanticPartialResponseReason(),
+            result.getSemanticPartialResponseType(), result.getDebugInfo(),
+            result.getSemanticQueryRewritesResultType());
     }
 
     /**